+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Copyright: Christian Holler
 - Added BoolValidator class.
 - Removed factory name '*apply*' from all value objects, failures and any other class. Naming the 
 factory is now only needed if it makes sense.
+- Added test cases for newly added code
 
 ## 1.0.0
 - Initial version.